Add VTX Menu to msposd (#8)
* setup vtxtable from msposd

* add vtxmenu

* populate vtxtable with supported freqs

* honor current frequency on startup

* store settings, use iwconfig

* move menu,add save&reboot option

* right selects action

* fix reading setings

* fix laggy ui

* add customizable vtxmenu

* fix save, add reboot

* add documentation

* cleanup

* use betflight style menu

* supoprt cutom commands

* disable mspvtx by default

* change needed after rebase

* support verbose flag

* hide keyboard input

* beatify commits

* fix compiler warnings

* float support

* add majestic restart command

* fix uppper case GUARD_INTERVAL

* add action response screen

* supoprt longer configs

* try fix bf menu crash

* add safeboot stick command

* safeguard channel 0 bug

* reboot fc during safeboot

* add more ini option

* add more options

* fixed small bug, merge marios ini

* add documentation

* last minute fixes

* fix regression for custom mesage

* remove fc reboot

* break MSP_CMD_STATUS

+++ bmp/region.h
@@ -15,7 +15,7 @@ extern "C"
     int create_region(int *handle, int x, int y, int width, int height);
     int prepare_bitmap(const char *filename, BITMAP *bitmap, int bFil, unsigned int u16FilColor, int enPixelFmt);
     int set_bitmap(int handle, BITMAP *bitmap);
-    int set_bitmapEx(int handle, BITMAP *bitmap, int BitsPerPixel);
+    unsigned long set_bitmapEx(int handle, BITMAP *bitmap, int BitsPerPixel);
     int unload_region(int *handle);
     void *get_directBMP(int handle);
 #ifdef __SIGMASTAR__
